+++ src/components/include/application_manager/policies/policy_handler_interface.h
@@ -492,7 +492,7 @@ class PolicyHandlerInterface {
       const std::string& policy_app_id,
       std::string& service_name,
       std::string& service_type,
-      std::vector<uint64_t>& handled_rpcs) const = 0;
+      std::vector<uint32_t>& handled_rpcs) const = 0;
 
 #ifdef EXTERNAL_PROPRIETARY_MODE
   /**
