+++ src/components/policy/policy_regular/include/policy/cache_manager_interface.h
@@ -235,7 +235,7 @@ class CacheManagerInterface {
       const std::string& policy_app_id,
       std::string& service_name,
       std::string& service_type,
-      std::vector<uint64_t>& handled_rpcs) const = 0;
+      std::vector<uint32_t>& handled_rpcs) const = 0;
 
   /**
    * @brief Allows to update 'vin' field in module_meta table.
